Pull ClusterConfig into its own file
